Fix #3672: Improve javalib java.net handling of NetworkInterface indices
